Rename RequestHandler to SessionFacade, move it into session package correct and extend the implementation
