+++ src/main/java/com/reka/lakatos/searchbyisbn/crawler/szechenyilibrary/session/exception/SessionActivationException.java
@@ -1,0 +1,7 @@
+package com.reka.lakatos.searchbyisbn.crawler.szechenyilibrary.session.exception;
+
+public class SessionActivationException extends RuntimeException {
+    public SessionActivationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
