+++ src/main/java/com/reka/lakatos/searchbyisbn/crawler/szechenyilibrary/session/exception/SessionDocumentException.java
@@ -1,0 +1,7 @@
+package com.reka.lakatos.searchbyisbn.crawler.szechenyilibrary.session.exception;
+
+public class SessionDocumentException extends RuntimeException {
+    public SessionDocumentException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
